pyproject.toml
[project]
name = "notify-bot"
version = "0.1.0"
description = "Minimal Telegram notification bot with HTTP API and CLI"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.10,<4",
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "pydantic>=2.0,<3",
    "python-dotenv>=1.0,<2",
    "httpx>=0.27,<1",
]

[project.scripts]
notify-bot = "notify_cli:main"
notify-bot-server = "notify_bot.main:run"

[tool.setuptools]
py-modules = ["notify_cli"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
